+++ R/mirai-package.R
@@ -50,9 +50,9 @@
 #'     (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
 #'
 #' @importFrom nanonext call_aio context cv cv_value is_error_value listen lock
-#'     mclock msleep opt parse_url pipe_notify recv recv_aio_signal request
-#'     request_signal send sha1 socket stat stop_aio unresolved wait weakref<-
-#' @importFrom stats runif
+#'     mclock msleep opt parse_url pipe_notify random recv recv_aio_signal
+#'     request request_signal send sha1 socket stat stop_aio unresolved wait
+#'     weakref<-
 #'
 #' @docType package
 #' @name mirai-package