+++ tests/tests.R
@@ -99,7 +99,12 @@ if (Sys.getenv("NOT_CRAN") == "true") {
   nanotest(is.matrix(status <- daemons()[["daemons"]]))
   nanotesto(status[, "online"])
   nanotesto(status[, "instance"])
-  nanotest(is.character(saisei(i = 1L)))
+  if (Sys.info()[["sysname"]] != "Windows") {
+    nanotest(is.character(saisei(i = 1L)))
+    nanotest(is.matrix(status <- daemons()[["daemons"]]))
+    nanotestz(status[, "online"])
+    nanotestz(status[, "instance"])
+  }
   nanotestz(daemons(0))
   Sys.sleep(1L)
 }